+++ tap/models.py
@@ -291,7 +291,7 @@ Index('tap_apiclicapara_authid', TapApiClientCustomPara.customauth_id)
 class TapApiAuth(Base):
     __tablename__ = 'tap_apiauth'
     id = Column(Integer, Sequence('seq_tapapiauth_id'), primary_key=True)
-    api_id = Column(Integer, ForeignKey('tap_api.id'))
+    api_id = Column(Integer, ForeignKey('tap_api.id'), nullable=False)
     api = relationship('TapApi', backref='auth_clients')
     client_id = Column(Integer, ForeignKey('tap_apiclient.id'), nullable=False)
     auth_client = relationship('TapApiClient', backref='auth_list')