+++ doc/source/CHANGES.rst
@@ -5,6 +5,7 @@
 0.3.5
 -----
 - Fix bug: can't get cache data in the page API->CONFIG->TAB[CACHE]
+- Config support: set api-config-cache time, default is 3 minutes, set ENVIRONMENT CONFIG TAP_API_RELOAD to override it.
 
 
 0.3.4